move ViewModelModule to correct package
