+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/app/AppComponent.kt
@@ -3,12 +3,17 @@ package com.lambdaschool.hackathon_portal.di.app
 import android.app.Application
 import com.lambdaschool.hackathon_portal.di.app.modules.AppModule
 import com.lambdaschool.hackathon_portal.di.activity.ActivityComponent
+import com.lambdaschool.hackathon_portal.di.app.modules.ViewModelsModule
 import dagger.BindsInstance
 import dagger.Component
 import javax.inject.Singleton
 
 @Singleton
-@Component(modules = [AppModule::class, ViewModelsModule::class])
+@Component(
+    modules = [
+        AppModule::class,
+        ViewModelsModule::class
+    ])
 interface AppComponent {
 
     fun getActivityComponentBuilder(): ActivityComponent.Builder
